requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
pillow
pdf2image
